Further extend shared codebase for integration tests
